Proposed fixes for pr #7464
- Simplified condition;
- Fixed test case;
- Test renamed and moved to TypeTest.cs;
