+++ mcs/class/corlib/corlib_test.dll.sources
@@ -30,7 +30,6 @@ System/IntegerTryParse.cs
 System/LazyTest.cs
 System/TupleTest.cs
 System/TypedReferenceTest.cs
-System/GetInterfaceTest.cs
 System.Collections/ArrayListTest.cs
 System.Collections/BitArrayTest.cs
 System.Collections/CaseInsensitiveComparerTest.cs
