+++ mcs/class/referencesource/mscorlib/system/rttype.cs
@@ -3413,14 +3413,11 @@ namespace System
 
 #if MONO
             List<RuntimeType> list = null;
+            var nameComparison = ignoreCase ? StringComparison.OrdinalIgnoreCase : StringComparison.Ordinal;
             foreach (RuntimeType t in GetInterfaces ()) {
 
-                if (ignoreCase) {
-                    if (t.Name.ToLowerInvariant() != name)
-                        continue;
-                } else {
-                    if (t.Name != name)
-                        continue;
+                if (!String.Equals(t.Name, name, nameComparison)) {
+                       continue;
                 }
 
                 if (list == null)
